requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==19.3.0
backcall==0.2.0
base58==2.0.1
bleach==3.2.1
blinker==1.4
boto3==1.15.10
botocore==1.18.10
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.1
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
Distance==0.1.3
entrypoints==0.3
enum-compat==0.0.3
fastapi==0.54.1
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
idna==2.10
importlib-metadata==1.6.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
jellyfish==0.8.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==8.2.0
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.19.2
packaging==20.3
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
py==1.8.1
pyarrow==1.0.1
pycld3==0.20
pycodestyle==2.5.0
pycparser==2.20
pydantic==1.4
pydeck==0.5.0b1
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==5.4.1
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
requests==2.24.0
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.14.0
starlette==0.13.2
streamlit==0.67.1
terminado==0.9.1
testpath==0.4.4
tinydb==4.1.1
toml==0.10.1
toolz==0.11.1
tornado==6.0.4
traitlets==4.3.3
tzlocal==2.1
urllib3==1.25.10
uvicorn==0.11.7
uvloop==0.14.0
validators==0.18.1
watchdog==0.10.3
wcwidth==0.1.9
webencodings==0.5.1
websockets==8.1
widgetsnbextension==3.5.1
zipp==3.1.0
